Update TMM protocol (#73)
* Update TMM protocol 

Change volumes used in making TMM and clarifying wording

* Update BHI-erm150 protocol

Add that erm should be dissolved in ethanol before adding to BHI media. We used to add the erm dry, but it takes to long to dissolve in the media. You have to dissolve the 150 mg erm in 1 ml ethanol first, and then add it to the media.